Lab3 3.2.1
